+++ ft_memset.c
@@ -6,15 +6,17 @@
 /*   By: kjurkows <kjurkows@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/15 20:43:16 by kjurkows          #+#    #+#             */
-/*   Updated: 2026/06/16 14:59:41 by kjurkows         ###   ########.fr       */
+/*   Updated: 2026/06/16 17:59:24 by kjurkows         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, long unsigned int n)
+#include "libft.h"
+
+void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned char		*ptr;
-	unsigned char		byte;
-	long unsigned int	i;
+	unsigned char	*ptr;
+	unsigned char	byte;
+	size_t			i;
 
 	i = 0;
 	byte = c;
