+++ ft_strlcpy.c
@@ -6,14 +6,15 @@
 /*   By: kjurkows <kjurkows@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/16 12:55:08 by kjurkows          #+#    #+#             */
-/*   Updated: 2026/06/16 14:22:45 by kjurkows         ###   ########.fr       */
+/*   Updated: 2026/06/16 17:56:17 by kjurkows         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long unsigned int	ft_strlcpy(char *dst, const char *src,
-	long unsigned int size)
+#include "libft.h"
+
+size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	long unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < size - 1 && src[i])
