+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: kjurkows <kjurkows@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/16 14:21:22 by kjurkows          #+#    #+#             */
-/*   Updated: 2026/06/16 14:29:29 by kjurkows         ###   ########.fr       */
+/*   Updated: 2026/06/16 17:56:27 by kjurkows         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, long unsigned int n)
+#include "libft.h"
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	long unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] && s1[i] == s2[i] && i < n)
